fix: Refactor global search validation [TECH-1656]
