+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/trackedentity/TrackedEntityOperationParamsMapper.java
@@ -27,6 +27,10 @@
  */
 package org.hisp.dhis.tracker.export.trackedentity;
 
+import static org.hisp.dhis.common.OrganisationUnitSelectionMode.ALL;
+import static org.hisp.dhis.common.OrganisationUnitSelectionMode.CHILDREN;
+import static org.hisp.dhis.common.OrganisationUnitSelectionMode.DESCENDANTS;
+import static org.hisp.dhis.common.OrganisationUnitSelectionMode.SELECTED;
 import static org.hisp.dhis.tracker.export.OperationsParamsValidator.validateOrgUnitMode;
 
 import java.util.ArrayList;
@@ -36,6 +40,7 @@ import java.util.Map;
 import java.util.Set;
 import javax.annotation.Nonnull;
 import lombok.RequiredArgsConstructor;
+import org.hisp.dhis.common.IllegalQueryException;
 import org.hisp.dhis.common.QueryFilter;
 import org.hisp.dhis.common.UID;
 import org.hisp.dhis.feedback.BadRequestException;
@@ -72,6 +77,10 @@ class TrackedEntityOperationParamsMapper {
 
   @Nonnull private final AclService aclService;
 
+  @Nonnull private final TrackedEntityStore trackedEntityStore;
+
+  @Nonnull private final TrackedEntityAttributeService trackedEntityAttributeService;
+
   @Transactional(readOnly = true)
   public TrackedEntityQueryParams map(TrackedEntityOperationParams operationParams)
       throws BadRequestException, ForbiddenException {
@@ -92,6 +101,9 @@ class TrackedEntityOperationParamsMapper {
     mapAttributeFilters(params, operationParams.getFilters());
 
     mapOrderParam(params, operationParams.getOrder());
+
+    validateTrackedEntityAttributeFilters(
+        program, trackedEntityType, operationParams, orgUnits, params);
 
     params
         .setProgram(program)
@@ -117,6 +129,8 @@ class TrackedEntityOperationParamsMapper {
         .setTrackedEntityUids(operationParams.getTrackedEntityUids())
         .setIncludeDeleted(operationParams.isIncludeDeleted())
         .setPotentialDuplicate(operationParams.getPotentialDuplicate());
+
+    validateGlobalSearchParameters(params);
 
     return params;
   }
@@ -281,6 +295,169 @@ class TrackedEntityOperationParamsMapper {
             "Cannot order by '"
                 + order.getField()
                 + "'. Tracked entities can be ordered by fields and tracked entity attributes.");
+      }
+    }
+  }
+
+  private void validateTrackedEntityAttributeFilters(
+      Program program,
+      TrackedEntityType trackedEntityType,
+      TrackedEntityOperationParams operationParams,
+      Set<OrganisationUnit> orgUnits,
+      TrackedEntityQueryParams params) {
+    if (program == null
+        && trackedEntityType == null
+        && operationParams.getFilters() != null
+        && orgUnits.isEmpty()) {
+      List<String> uniqueAttributeIds =
+          trackedEntityAttributeService.getAllSystemWideUniqueTrackedEntityAttributes().stream()
+              .map(TrackedEntityAttribute::getUid)
+              .toList();
+
+      for (String att : params.getFilterIds()) {
+        if (!uniqueAttributeIds.contains(att)) {
+          throw new IllegalQueryException(
+              "Either a program or tracked entity type must be specified");
+        }
+      }
+    }
+  }
+
+  private void validateGlobalSearchParameters(TrackedEntityQueryParams params)
+      throws IllegalQueryException {
+    if (!isLocalSearch(params, params.getUser())) {
+
+      if (params.hasFilters()) {
+        List<String> searchableAttributeIds = getSearchableAttributeIds(params);
+        validateSearchableAttributes(params, searchableAttributeIds);
+      }
+
+      int maxTeiLimit = getMaxTeiLimit(params);
+      checkIfMaxTeiLimitIsReached(params, maxTeiLimit);
+      params.setMaxTeLimit(maxTeiLimit);
+    }
+  }
+
+  private List<String> getSearchableAttributeIds(TrackedEntityQueryParams params) {
+    List<String> searchableAttributeIds = new ArrayList<>();
+
+    if (params.hasProgram()) {
+      searchableAttributeIds.addAll(params.getProgram().getSearchableAttributeIds());
+    }
+
+    if (params.hasTrackedEntityType()) {
+      searchableAttributeIds.addAll(params.getTrackedEntityType().getSearchableAttributeIds());
+    }
+
+    if (!params.hasProgram() && !params.hasTrackedEntityType()) {
+      searchableAttributeIds.addAll(
+          trackedEntityAttributeService.getAllSystemWideUniqueTrackedEntityAttributes().stream()
+              .map(TrackedEntityAttribute::getUid)
+              .toList());
+    }
+
+    return searchableAttributeIds;
+  }
+
+  private void validateSearchableAttributes(
+      TrackedEntityQueryParams params, List<String> searchableAttributeIds) {
+    List<String> violatingAttributes = new ArrayList<>();
+
+    for (String attributeId : params.getFilterIds()) {
+      if (!searchableAttributeIds.contains(attributeId)) {
+        violatingAttributes.add(attributeId);
+      }
+    }
+
+    if (!violatingAttributes.isEmpty()) {
+      throw new IllegalQueryException(
+          "Non-searchable attribute(s) can not be used during global search:  "
+              + violatingAttributes);
+    }
+  }
+
+  private int getMaxTeiLimit(TrackedEntityQueryParams params) {
+    int maxTeiLimit = 0;
+    if (params.hasTrackedEntityType()) {
+      maxTeiLimit = params.getTrackedEntityType().getMaxTeiCountToReturn();
+
+      if (!params.hasTrackedEntities() && isTeTypeMinAttributesViolated(params)) {
+        throw new IllegalQueryException(
+            "At least "
+                + params.getTrackedEntityType().getMinAttributesRequiredToSearch()
+                + " attributes should be mentioned in the search criteria.");
+      }
+    }
+
+    if (params.hasProgram()) {
+      maxTeiLimit = params.getProgram().getMaxTeiCountToReturn();
+
+      if (!params.hasTrackedEntities() && isProgramMinAttributesViolated(params)) {
+        throw new IllegalQueryException(
+            "At least "
+                + params.getProgram().getMinAttributesRequiredToSearch()
+                + " attributes should be mentioned in the search criteria.");
+      }
+    }
+
+    return maxTeiLimit;
+  }
+
+  private boolean isLocalSearch(TrackedEntityQueryParams params, User user) {
+    Set<OrganisationUnit> localOrgUnits = user.getOrganisationUnits();
+
+    Set<OrganisationUnit> searchOrgUnits = new HashSet<>();
+
+    if (params.isOrganisationUnitMode(SELECTED)) {
+      searchOrgUnits = params.getOrgUnits();
+    } else if (params.isOrganisationUnitMode(CHILDREN)
+        || params.isOrganisationUnitMode(DESCENDANTS)) {
+      for (OrganisationUnit orgUnit : params.getOrgUnits()) {
+        searchOrgUnits.addAll(orgUnit.getChildren());
+      }
+    } else if (params.isOrganisationUnitMode(ALL)) {
+      searchOrgUnits.addAll(organisationUnitService.getRootOrganisationUnits());
+    } else {
+      searchOrgUnits.addAll(user.getTeiSearchOrganisationUnitsWithFallback());
+    }
+
+    for (OrganisationUnit ou : searchOrgUnits) {
+      if (!ou.isDescendant(localOrgUnits)) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  private boolean isTeTypeMinAttributesViolated(TrackedEntityQueryParams params) {
+    if (params.hasUniqueFilter()) {
+      return false;
+    }
+
+    return (!params.hasFilters()
+            && params.getTrackedEntityType().getMinAttributesRequiredToSearch() > 0)
+        || (params.hasFilters()
+            && params.getFilters().size()
+                < params.getTrackedEntityType().getMinAttributesRequiredToSearch());
+  }
+
+  private boolean isProgramMinAttributesViolated(TrackedEntityQueryParams params) {
+    if (params.hasUniqueFilter()) {
+      return false;
+    }
+
+    return (!params.hasFilters() && params.getProgram().getMinAttributesRequiredToSearch() > 0)
+        || (params.hasFilters()
+            && params.getFilters().size() < params.getProgram().getMinAttributesRequiredToSearch());
+  }
+
+  private void checkIfMaxTeiLimitIsReached(TrackedEntityQueryParams params, int maxTeiLimit) {
+    if (maxTeiLimit > 0) {
+      int teCount = trackedEntityStore.getTrackedEntityCountWithMaxTrackedEntityLimit(params);
+
+      if (teCount > maxTeiLimit) {
+        throw new IllegalQueryException("maxteicountreached");
       }
     }
   }
